@@ -15,6 +15,7 @@
     "@fortawesome/fontawesome-svg-core": "^6.6.0",
     "@fortawesome/free-solid-svg-icons": "^6.6.0",
     "@fortawesome/vue-fontawesome": "^3.0.8",
+    "@vue/reactivity": "^3.5.12",
     "axios": "^1.7.7",
     "bootstrap": "^5.3.3",
     "chart.js": "^4.4.4",
@@ -24,9 +25,10 @@
     "jspdf": "^2.5.2",
     "pinia": "^2.1.7",
     "popper.js": "^1.16.1",
-    "vue": "^3.4.29",
+    "vue": "^3.5.12",
     "vue-router": "^4.3.3",
-    "vuex": "^4.1.0"
+    "vuex": "^4.1.0",
+    "watch": "^1.0.2"
   },
   "devDependencies": {
     "@rushstack/eslint-patch": "^1.8.0",
