@@ -5,10 +5,10 @@
                 <tr>
                     <th>Order Id</th>
                     <th>Customer Name</th>
+                    <th>Phone</th>
                     <th>Restaurant</th>
-                    <th>Date</th>
-                    <th>Status</th>
-                    <th>Price</th>
+                    <th>Total Price</th>
+                    <th>Food Commission</th>
                     <th>View Details</th>
                 </tr>
             </thead>
@@ -16,15 +16,13 @@
                 <transition-group name="fade" mode="out-in">
                     <tr v-for="order in filteredOrders" :key="order.id">
                         <td>{{ order.id }}</td>
-                        <td>{{ order.customerName }}</td>
-                        <td>{{ order.restaurant }}</td>
-                        <td>{{ formatDate(order.date) }}</td>
-                        <td :class="`status-${order.status.toLowerCase().replace(' ', '-')}`">
-                            <strong>{{ order.status }}</strong>
-                        </td>
-                        <td>{{ formatCurrency(order.price) }}</td>
+                        <td>{{ order.name }}</td>
+                        <td>{{ order.phone }}</td>
+                        <td>{{ order.resturant }}</td>
+                        <td>{{ formatCurrency(order.totalPrice) }}</td>
+                        <td>{{ formatCurrency(order.foodCommission) }}</td>
                         <td>
-                            <button class="btn btn-details" @click="viewDetails(order)">
+                            <button class="btn btn-details" @click="openModal(order)">
                                 <i class="fas fa-eye"></i> View Details
                             </button>
                         </td>
@@ -32,166 +30,241 @@
                 </transition-group>
             </tbody>
         </table>
+
+        <!-- Receipt Modal -->
+        <div class="modal fade" id="orderModal" tabindex="-1" aria-labelledby="orderModalLabel" aria-hidden="true">
+            <div class="modal-dialog modal-lg">
+                <div class="modal-content">
+                    <div class="modal-header">
+                        <h5 class="modal-title" id="orderModalLabel">Order Receipt</h5>
+                        <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+                    </div>
+                    <div class="modal-body">
+                        <div class="receipt" ref="receiptContent">
+                            <h5 class="text-center">Receipt</h5>
+                            <hr />
+                            <div class="row">
+                                <div class="col-6">
+                                    <p><strong>Order Number:</strong> {{ orderDetails.id }}</p>
+                                    <p><strong>Date:</strong> {{ orderDetails.date }}</p>
+                                    <p><strong>Time:</strong> {{ orderDetails.time }}</p>
+                                </div>
+                                <div class="col-6 text-end">
+                                    <p><strong>Customer Name:</strong> {{ orderDetails.name }}</p>
+                                    <p><strong>Phone Number:</strong> {{ orderDetails.phone }}</p>
+                                    <p><strong>Address:</strong> {{ orderDetails.address }}</p>
+                                </div>
+                            </div>
+                            <hr />
+                            <table class="table">
+                                <thead>
+                                    <tr>
+                                        <th scope="col">Description</th>
+                                        <th scope="col" class="text-end">Amount (PKR)</th>
+                                    </tr>
+                                </thead>
+                                <tbody>
+                                    <tr>
+                                        <td>Total Product Price</td>
+                                        <td class="text-end">{{ orderDetails.totalPrice }} PKR</td>
+                                    </tr>
+                                    <tr>
+                                        <td>Delivery Charges</td>
+                                        <td class="text-end">200 PKR</td>
+                                    </tr>
+                                    <tr>
+                                        <td>GST (16%)</td>
+                                        <td class="text-end">{{ gstAmount }} PKR</td>
+                                    </tr>
+                                    <tr>
+                                        <td>Food Pro Commission</td>
+                                        <td class="text-end">{{ orderDetails.foodCommission }} PKR</td>
+                                    </tr>
+                                    <tr class="table-primary">
+                                        <td><strong>Total Restaurant Got</strong></td>
+                                        <td class="text-end"><strong>{{ totalRestaurantGot }} PKR</strong></td>
+                                    </tr>
+                                    <tr class="table-success">
+                                        <td><strong>Profit</strong></td>
+                                        <td class="text-end"><strong>{{ profit }} PKR</strong></td>
+                                    </tr>
+                                </tbody>
+                            </table>
+                        </div>
+                    </div>
+                    <div class="modal-footer">
+                        <button type="button" class="btn btn-danger" data-bs-dismiss="modal">Close</button>
+                        <button type="button" class="btn btn-success" @click="downloadReceiptAsPDF">Download Receipt</button>
+                    </div>
+                </div>
+            </div>
+        </div>
     </div>
 </template>
 
 <script setup>
-import { defineProps } from 'vue';
+import { ref, defineProps, computed } from 'vue';
+import jsPDF from 'jspdf';
+import html2canvas from 'html2canvas';
 
+// Accept filteredOrders as a prop
 const props = defineProps({
     filteredOrders: {
         type: Array,
-        required: true,
-    },
+        required: true
+    }
 });
 
-// Format date function
-const formatDate = (dateString) => {
-    const options = { year: 'numeric', month: 'short', day: 'numeric' };
-    return new Date(dateString).toLocaleDateString(undefined, options);
+// Format currency function
+const formatCurrency = (amount) => {
+    return `Rs ${amount.toFixed(2)}`;
 };
 
-// Format currency function
-const formatCurrency = (price) => {
-    return `Rs ${price.toFixed(2)}`; 
-};
+// Order details ref
+const orderDetails = ref({});
 
 // View details function
-const viewDetails = (order) => {
-    console.log('Viewing details for:', order);
+const openModal = (order) => {
+    const currentDate = new Date();
+    const formattedDate = currentDate.toLocaleDateString();
+    const formattedTime = currentDate.toLocaleTimeString();
+
+    orderDetails.value = {
+        ...order,
+        date: formattedDate,
+        time: formattedTime,
+        address: order.address || 'N/A' // Assuming address is in order
+    };
+    const modal = new bootstrap.Modal(document.getElementById("orderModal"));
+    modal.show();
+};
+
+// Calculate GST
+const gstAmount = computed(() => {
+    return (orderDetails.value.totalPrice * 0.16).toFixed(2);
+});
+
+// Calculate total amount restaurant got
+const totalRestaurantGot = computed(() => {
+    return (orderDetails.value.totalPrice - gstAmount.value - orderDetails.value.foodCommission + 200).toFixed(2);
+});
+
+// Calculate profit
+const profit = computed(() => {
+    return (orderDetails.value.foodCommission / 2).toFixed(2);
+});
+
+// Download receipt as PDF
+const downloadReceiptAsPDF = async () => {
+    const receiptElement = document.querySelector('.receipt'); // Use a query selector to get the receipt content
+
+    try {
+        // Use html2canvas to take a snapshot of the receipt content
+        const canvas = await html2canvas(receiptElement, { scale: 2 });
+        const imgData = canvas.toDataURL('image/png');
+
+        // Create a PDF using jsPDF
+        const pdf = new jsPDF('p', 'mm', 'a4');
+        const pageWidth = pdf.internal.pageSize.getWidth();
+        const imgWidth = pageWidth - 20; // Leave some margin
+        const imgHeight = (canvas.height * imgWidth) / canvas.width; // Maintain aspect ratio
+
+        pdf.addImage(imgData, 'PNG', 10, 10, imgWidth, imgHeight);
+        
+        // Save the PDF with the order ID
+        pdf.save(`Receipt_${orderDetails.value.id}.pdf`);
+    } catch (error) {
+        console.error('Error generating PDF:', error);
+        alert('Failed to generate PDF. Please try again later.');
+    }
 };
 </script>
 
 <style scoped>
 .table-responsive {
     max-height: 500px;
-    /* Increased height for better visibility */
     overflow-y: auto;
-    /* Add vertical scrolling */
     margin: 20px;
-    /* Margin around the table */
     padding: 40px;
     border-radius: 10px;
-    /* Rounded corners */
     background-color: #fdfdfd;
-    
 }
 
 .table {
     width: 100%;
-    /* Full width */
     border-collapse: collapse;
-    /* Remove double borders */
 }
 
 thead th {
     background-color: #00754A;
-    /* Soft blue header background */
     color: #ffffff;
-    /* Header text color */
     padding: 15px;
-    /* Padding for better spacing */
     text-align: left;
-    /* Align text to left */
     font-weight: 600;
-    /* Slightly bolder header text */
     border-bottom: 2px solid #00754A;
-    /* Stronger border for emphasis */
 }
 
 tbody tr {
     transition: background-color 0.2s;
-    /* Smooth transition for hover effect */
 }
 
 tbody tr:hover {
     background-color: #e8f0fe;
-    /* Light blue on hover */
 }
 
 table td {
     padding: 12px;
-    /* Padding for cells */
     border-bottom: 1px solid #ddd;
-    /* Light border between rows */
     vertical-align: middle;
-    /* Center align cells vertically */
     font-size: 14px;
-    /* Base font size */
-}
-
-.status-in-progress {
-    color: #ffa000;
-    /* Orange for in-progress */
-}
-
-.status-completed {
-    color: #388e3c;
-    /* Green for completed */
-}
-
-.status-cancelled {
-    color: #d32f2f;
-    /* Red for cancelled */
 }
 
 .btn-details {
     display: inline-flex;
     align-items: center;
     background-color: #00754A;
-    /* Button color */
     color: white;
-    /* Button text color */
     border: none;
     border-radius: 5px;
     padding: 8px 12px;
-    /* Button padding */
     cursor: pointer;
     transition: background-color 0.3s, transform 0.2s;
-    /* Transition for hover */
     font-size: 14px;
-    /* Button font size */
 }
 
 .btn-details i {
     margin-right: 5px;
-    /* Space between icon and text */
 }
 
 .btn-details:hover {
     background-color: #00754adc;
-    /* Darker blue on hover */
     transform: scale(1.05);
-    /* Scale up button on hover */
+}
+
+/* Receipt styles */
+.receipt {
+    padding: 20px;
+    border: 1px solid #ddd;
+    box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.1);
+    border-radius: 5px;
+}
+
+.receipt h5 {
+    margin-bottom: 20px;
 }
 
 /* Responsive adjustments */
 @media (max-width: 768px) {
     table {
         font-size: 12px;
-        /* Smaller font for mobile */
     }
 
     .btn-details {
-        font-size: 12px;
-        /* Smaller button font */
+        padding: 6px 10px;
     }
 
-    thead th,
-    tbody td {
+    .receipt {
         padding: 10px;
-        /* Smaller padding for compactness */
     }
-}
-
-/* Animation styles */
-.fade-enter-active,
-.fade-leave-active {
-    transition: opacity 0.5s;
-}
-
-.fade-enter,
-.fade-leave-to {
-    opacity: 0;
 }
 </style>
