@@ -1,15 +1,15 @@
 <template>
+
   <div class="container my-5">
     <div class="row">
       <!-- Left side: Delivery Address Section -->
       <div class="col-lg-7 mb-4">
         <div class="card p-4">
           <h3>Delivery address</h3>
-          <router-link to="/orderreq"><a class="nxt-btn">Order Request</a></router-link>
-          <div class="map-container mb-3 mt-2">
-
-            <img src="../../assets/images/Screenshot 2024-10-07 151047.png" alt="Map" class="img-fluid rounded">
-          </div>
+          <router-link to="/orderreq">
+            <a class="nxt-btn">Order Request</a>
+          </router-link>
+          <div class="map-container mb-3 mt-2"></div>
           <div>
             <i class="fa-solid fa-location-crosshairs"></i>
             <p class="address">
@@ -18,16 +18,18 @@
               Lahore
             </p>
           </div>
-          <!--  -->
           <div class="mt-3">
-            <textarea class="form-control" rows="3" placeholder="Note to rider - e.g., building, landmark"></textarea>
+            <textarea
+              class="form-control"
+              rows="3"
+              placeholder="Note to rider - e.g., building, landmark"
+            ></textarea>
           </div>
-          <!--  -->
-          <hr>
+          <hr />
           <p class="contactless-toggle">
             Contactless delivery
             <label class="switch">
-              <input type="checkbox">
+              <input type="checkbox" />
               <span class="slider round"></span>
             </label>
           </p>
@@ -63,7 +65,6 @@
               <p class="d-flex justify-content-between">VAT <span>Rs. 0</span></p>
             </div>
 
-
             <div>
               <div class="d-flex justify-content-between fw-bold">
                 <p class="total">Total</p>
@@ -71,7 +72,6 @@
               </div>
               <span class="total-inc">(Incl. VAT)</span>
             </div>
-
           </div>
         </div>
       </div>
@@ -84,12 +84,9 @@
         <div class="card mb-4">
           <div class="card-body">
             <h4 class="card-title">Delivery options</h4>
-            <!-- Delivery options content -->
             <div class="form-check mt-3 delivery-check">
-              <input class="form-check-input" type="radio" name="deliveryOption" id="standard" checked>
-              <label class="form-check-label" for="standard">
-                Standard 15 - 30 mins
-              </label>
+              <input class="form-check-input" type="radio" name="deliveryOption" id="standard" checked />
+              <label class="form-check-label" for="standard">Standard 15 - 30 mins</label>
             </div>
           </div>
         </div>
@@ -110,44 +107,39 @@
             <h4 class="card-title">Payment</h4>
             <p>Available payment methods</p>
             <div class="form-check delivery-check">
-              <input class="form-check-input" type="radio" name="paymentMethod" id="easypaisa" checked>
-              <label class="form-check-label" for="easypaisa">
-                Cash On Delivery (COD)
-              </label>
+              <input class="form-check-input" type="radio" name="paymentMethod" id="easypaisa" checked />
+              <label class="form-check-label" for="easypaisa">Cash On Delivery (COD)</label>
             </div>
           </div>
         </div>
 
-        <!-- Payment Card -->
-        <div>
-          <!-- Save and Continue Button -->
-          <div class="mt-4">
-            <button class="btn btn-green w-100">Place Order</button>
-          </div>
+        <!-- Save and Continue Button -->
+        <div class="mt-4">
+          <button class="btn btn-green w-100">Place Order</button>
         </div>
         <div class="general-terms">
           <p>By making this purchase you agree to our terms and conditions.</p>
-          <p>I agree that placing the order places me under an obligation to make a payment in accordance with the
-            General Terms and Conditions.</p>
+          <p>I agree that placing the order places me under an obligation to make a payment in accordance with the General Terms and Conditions.</p>
         </div>
-
       </div>
 
     </div>
   </div>
 </template>
 
-<script>
-export default {
-  name: "CheckOut",
-};
+
+<script setup>
+import { defineComponent } from 'vue';
+
+defineComponent({
+  name: 'CheckOut',
+});
 </script>
 
 <style scoped>
 .card {
   border: none;
 }
-
 .card h3,
 h4 {
   font-family: 'Agrandir', 'Open Sans', 'Helvetica Neue', sans-serif;
@@ -157,29 +149,24 @@ h4 {
 .my-5 {
   padding: 60px 80px 60px 80px;
 }
-
 .map-container img {
   width: 100%;
   height: auto;
 }
-
 .fa-location-crosshairs,
 .address {
   display: inline-block;
 }
-
 .fa-location-crosshairs {
   position: relative;
   bottom: 20px;
   margin: 10px;
   font-size: 20px;
 }
-
 .total {
   font-size: 1.5rem;
   margin-bottom: -5px;
 }
-
 .total-inc {
   font-size: 14px;
 }
@@ -215,7 +202,7 @@ h4 {
 
 .slider:before {
   position: absolute;
-  content: "";
+  content: '';
   height: 12px;
   width: 12px;
   left: 4px;
@@ -264,11 +251,12 @@ input:checked+.slider:before {
   border-radius: 10px;
 }
 
+
 .general-terms {
   margin-top: 20px;
   color: rgb(97, 97, 97);
   font-size: 12px;
 }
 
-/**/
 </style>
+
