@@ -60,7 +60,9 @@
     </nav>
 </template>
 
-<script>
+
+<script setup>
+
 import { ref, onMounted, onBeforeUnmount } from 'vue';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 import { library } from '@fortawesome/fontawesome-svg-core';
@@ -74,6 +76,7 @@ import {
     faHeart, // Added heart icon
 } from '@fortawesome/free-solid-svg-icons';
 
+// Add icons to the library
 library.add(
     faShoppingCart,
     faUtensils,
@@ -84,45 +87,35 @@ library.add(
     faHeart // Added heart icon to library
 );
 
-export default {
-    name: 'LoginHeader',
-    components: {
-        FontAwesomeIcon,
-    },
-    setup() {
-        const dropdownOpen = ref(false);
-        const username = ref('Guest');
-        const dropdown = ref(null);
+// Reactive state
+const dropdownOpen = ref(false);
+const username = ref('Guest');
+const dropdown = ref(null);
 
-        const toggleDropdown = (event) => {
-            event.stopPropagation(); // Prevent bubbling to the document click listener
-            dropdownOpen.value = !dropdownOpen.value;
-        };
-
-        const closeDropdown = (event) => {
-            // Only close if the click is outside the dropdown
-            if (dropdown.value && !dropdown.value.contains(event.target)) {
-                dropdownOpen.value = false;
-            }
-        };
-
-        onMounted(() => {
-            document.addEventListener('click', closeDropdown);
-        });
-
-        onBeforeUnmount(() => {
-            document.removeEventListener('click', closeDropdown);
-        });
-
-        return {
-            dropdownOpen,
-            username,
-            toggleDropdown,
-            dropdown,
-        };
-    },
+// Toggle dropdown visibility
+const toggleDropdown = (event) => {
+    event.stopPropagation(); // Prevent bubbling to the document click listener
+    dropdownOpen.value = !dropdownOpen.value;
 };
+
+// Close dropdown if clicked outside
+const closeDropdown = (event) => {
+    if (dropdown.value && !dropdown.value.contains(event.target)) {
+        dropdownOpen.value = false;
+    }
+};
+
+
+// Lifecycle hooks
+onMounted(() => {
+    document.addEventListener('click', closeDropdown);
+});
+
+onBeforeUnmount(() => {
+    document.removeEventListener('click', closeDropdown);
+});
 </script>
+
 
 <style scoped>
 .navbar {
@@ -170,11 +163,6 @@ export default {
     margin-right: 5px;
 }
 
-.arrow {
-    font-size: 16px;
-    margin-left: 5px;
-}
-
 .dropdownmenu {
     position: absolute;
     top: 100%;
@@ -185,12 +173,8 @@ export default {
     border-radius: 8px;
     min-width: 200px;
     z-index: 1000;
-}
 
-[v-show="true"] .dropdownmenu {
-    display: block;
-    transform: translateY(0);
-    opacity: 1;
+
 }
 
 .dropdownmenu ul {
@@ -213,12 +197,9 @@ export default {
     border-radius: 5px;
 }
 
-.dropdownmenu li i {
-    margin-right: 10px;
-    font-size: 18px;
-}
 
-.favorites-button {
+.cart-icon {
+
     margin-left: 20px;
     display: flex;
     align-items: center;
@@ -250,9 +231,5 @@ export default {
 
 .cart-icon-button:hover {
     color: #00754a;
-}
-
-.icon {
-    font-size: 20px;
 }
 </style>
