@@ -1,6 +1,5 @@
 <template>
     <div>
-      
         <!-- Main Navbar -->
         <nav class="navbar">
             <div class="navbar-left">
@@ -10,50 +9,40 @@
             </div>
 
             <div class="navbar-right">
-                <!-- Log In  buttons -->
-                <p>Already have Account ? <button class="auth-button login" @click="goToLogin">Back</button></p>
+
+                <!-- Log In buttons -->
+                <button class="auth-button login" @click="goToLogin">Log in</button>
+
             </div>
         </nav>
     </div>
 </template>
 
-<script>
+<script setup>
 import { ref } from 'vue';
 import { FontAwesomeIcon } from '@fortawesome/vue-fontawesome';
 import { useRouter } from 'vue-router';
 
-export default {
-    name: 'LoginHeader',
-    components: {
-        FontAwesomeIcon,
-    },
-    setup() {
-        const router = useRouter();
-        const showBanner = ref(true);
-
-        // Function to close the business banner
-        const closeBanner = () => {
-            showBanner.value = false;
-        };
-
-        // Navigation functions
-        const goToLogin = () => {
-            router.push('/');
-        };
+// Router instance
+const router = useRouter();
 
 
-        return {
-            closeBanner,
-            goToLogin,
-            
-        };
-    },
+// State management
+const showBanner = ref(true);
+
+
+// Function to close the business banner
+const closeBanner = () => {
+    showBanner.value = false;
+};
+
+// Navigation function
+const goToLogin = () => {
+    router.push('/login');
 };
 </script>
 
 <style scoped>
-
-
 .close-icon {
     cursor: pointer;
     font-size: 20px;
@@ -109,5 +98,4 @@ export default {
     background-color: #2e8d6a;
     color: white;
 }
-
-</style>  
+</style>
