@@ -8,7 +8,7 @@
         <span class="close" @click="closeModal">&times;</span>
         <h3 v-if="!selectedOrder">New Orders</h3>
         <h3 v-if="selectedOrder">Order Details</h3>
-        
+
         <div v-if="!selectedOrder" class="order-requests">
           <div v-for="(order, index) in orders" :key="index" class="order-card">
             <div class="order-details">
@@ -57,131 +57,129 @@
   </div>
 </template>
 
-    
-    <script>
-    export default {
-      data() {
-        return {
-          showModal: false,
-          selectedOrder: null, // Store the currently viewed order
-          orders: [], // List of new orders
-          acceptedOrders: [],
-          rejectedOrders: [],
-          interval: null,
-        };
-      },
-      mounted() {
-        this.populateDummyOrders();
-      },
-      methods: {
-        populateDummyOrders() {
-          // Generate some dummy orders
-          for (let i = 1; i <= 5; i++) {
-            this.orders.push({
-              id: `#0${i}`,
-              username: `User ${i}`,
-              items: [`Item ${i}a`, `Item ${i}b`],
-              subtotal: (i * 100).toFixed(2),
-              total: (i * 120).toFixed(2),
-              timer: 3000, // 3000 seconds timer
-            });
-          }
-          this.startTimers();
-        },
-        startTimers() {
-          this.orders.forEach((order) => {
-            const orderInterval = setInterval(() => {
-              if (order.timer > 0) {
-                order.timer--;
-              } else {
-                this.rejectOrder(order);
-                clearInterval(orderInterval); // Stop the timer if time runs out
-              }
-            }, 1000);
-          });
-        },
-        acceptOrder(order) {
-          this.acceptedOrders.push(order);
-          this.orders = this.orders.filter((o) => o.id !== order.id);
-          this.playAlertSound(); // Play sound when an order is accepted
-        },
-        rejectOrder(order) {
-          this.rejectedOrders.push(order);
-          this.orders = this.orders.filter((o) => o.id !== order.id);
-          this.playAlertSound(); // Play sound when an order is rejected
-        },
-        viewOrder(order) {
-          this.selectedOrder = order; // Set the selected order to show details
-        },
-        clearSelectedOrder() {
-          this.selectedOrder = null; // Clear selected order
-        },
-        closeModal() {
-          this.showModal = false; 
-          this.clearSelectedOrder(); // Clear selection when closing modal
-        },
-        playAlertSound() {
-          const audio = new Audio(new URL('@/assets/sounds/order_ringtone.wav', import.meta.url).href);
-          audio.loop = true; // Set to loop
-          audio.play().catch((error) => {
-            console.error('Error playing audio:', error);
-          });
-        },
-      },
-    };
-    </script>
-    
-    
-  
-    <style scoped>
-    .main-cont {
-      padding-top: 70px;
-      height: 100vh;
-      position: relative;
-    }
-    
-    .modal {
-      position: fixed;
-      z-index: 100;
-      left: 0;
-      top: 0;
-      background-color: rgba(0, 0, 0, 0.2);
-      display: flex;
-      justify-content: end;
-      align-items: center;
-    }
-    
-    .modal-content {
-      background-color: white;
-      padding: 20px;
-      border-radius: 8px;
-      position: relative;
-      max-height: 400px;
-      max-width: 500px;
-      border: none;
-      
-    }
-    
-    .close {
-      position: absolute;
-      top: 10px;
-      right: 15px;
-      font-size: 20px;
-      cursor: pointer;
-    }
-    
-    .order-card {
-      border: 1px solid #e0e0e0;
-      margin-bottom: 10px;
-      padding: 10px;
-      border-radius: 5px;
-    }
-    .order-requests{
-        overflow-y: auto;
-    }
-    .order-actions button {
-      margin-right: 5px;
-    }
-    </style>
-    
-  
+<script setup>
+import { ref, onMounted } from 'vue';
+
+const showModal = ref(false);
+const selectedOrder = ref(null); // Store the currently viewed order
+const orders = ref([]); // List of new orders
+const acceptedOrders = ref([]);
+const rejectedOrders = ref([]);
+const interval = ref(null);
+
+const populateDummyOrders = () => {
+  // Generate some dummy orders
+  for (let i = 1; i <= 5; i++) {
+    orders.value.push({
+      id: `#0${i}`,
+      username: `User ${i}`,
+      items: [`Item ${i}a`, `Item ${i}b`],
+      subtotal: (i * 100).toFixed(2),
+      total: (i * 120).toFixed(2),
+      timer: 3000, // 3000 seconds timer
+    });
+  }
+  startTimers();
+};
+
+const startTimers = () => {
+  orders.value.forEach((order) => {
+    const orderInterval = setInterval(() => {
+      if (order.timer > 0) {
+        order.timer--;
+      } else {
+        rejectOrder(order);
+        clearInterval(orderInterval); // Stop the timer if time runs out
+      }
+    }, 1000);
+  });
+};
+
+const acceptOrder = (order) => {
+  acceptedOrders.value.push(order);
+  orders.value = orders.value.filter((o) => o.id !== order.id);
+  playAlertSound(); // Play sound when an order is accepted
+};
+
+const rejectOrder = (order) => {
+  rejectedOrders.value.push(order);
+  orders.value = orders.value.filter((o) => o.id !== order.id);
+  playAlertSound(); // Play sound when an order is rejected
+};
+
+const viewOrder = (order) => {
+  selectedOrder.value = order; // Set the selected order to show details
+};
+
+const clearSelectedOrder = () => {
+  selectedOrder.value = null; // Clear selected order
+};
+
+const closeModal = () => {
+  showModal.value = false; 
+  clearSelectedOrder(); // Clear selection when closing modal
+};
+
+const playAlertSound = () => {
+  const audio = new Audio(new URL('@/assets/sounds/order_ringtone.wav', import.meta.url).href);
+  audio.loop = true; // Set to loop
+  audio.play().catch((error) => {
+    console.error('Error playing audio:', error);
+  });
+};
+
+// Lifecycle hook to populate orders on mount
+onMounted(() => {
+  populateDummyOrders();
+});
+</script>
+
+<style scoped>
+.main-cont {
+  padding-top: 70px;
+  height: 100vh;
+  position: relative;
+}
+
+.modal {
+  position: fixed;
+  z-index: 100;
+  left: 0;
+  top: 0;
+  background-color: rgba(0, 0, 0, 0.2);
+  display: flex;
+  justify-content: end;
+  align-items: center;
+}
+
+.modal-content {
+  background-color: white;
+  padding: 20px;
+  border-radius: 8px;
+  position: relative;
+  max-height: 400px;
+  max-width: 500px;
+  border: none;
+}
+
+.close {
+  position: absolute;
+  top: 10px;
+  right: 15px;
+  font-size: 20px;
+  cursor: pointer;
+}
+
+.order-card {
+  border: 1px solid #e0e0e0;
+  margin-bottom: 10px;
+  padding: 10px;
+  border-radius: 5px;
+}
+.order-requests {
+  overflow-y: auto;
+}
+.order-actions button {
+  margin-right: 5px;
+}
+</style>
