@@ -17,6 +17,8 @@ import Main_landing from '../views/Login_Signup/Main_Landing.vue'
 import Business_Landing from '../views/Login_Signup/Bussiness_Landing.vue'
 import AdminFinanceDashboard from '../components/Admin/AdminFinanceDashboard.vue'
 import RestaurantOwner_Dashboard from '../views/RestauranOnwer/RestaurantOwner_Dashboard.vue'
+import CategoryNavbar from '../components/Customer/CategoryNavbar.vue'
+import AddtoCart from '../components/AddtoCart.vue'
 import ResturantOwner from '@/components/RestaurantOwner/ResturantOwner.vue'
 import RestaurantCard from '@/components/Customer/RestaurantCard.vue'
 import adminRoutes from './adminRoutes';
@@ -32,10 +34,8 @@ import hello from '@/views/hello.vue'
 // import SeeReviews from '../components/Customer/SeeReviews.vue'
 
 const routes = [
-  ...adminRoutes,
-  ...customerRoutes,
-  ...restaurantOwnerRoutes,
-  ...otherRoutes,
+
+
   {
     path: '/Navbarheader',
     name: 'Navbarheader',
@@ -50,8 +50,18 @@ const routes = [
         path: '/orderreq',
         name: 'orderreq',
         component: OrderReq
-      }
+      },
+      {
+        path: '/caart',
+        name: 'addtocard',
+        component: AddtoCart
+      },
     ]
+  },
+  {
+    path: '/catnvbr',
+    name: 'categorynavbar',
+    component: CategoryNavbar
   },
   {
     path: '/RestaurantPage',
@@ -59,43 +69,10 @@ const routes = [
     component: Restaurantpage
   },
   {
-    path: '/reset-password',
-        name: 'SetNewPassword',
-        component: hello
-  },
-  {
-    path: '/orderrating',
-    name: 'OrderRating',
-    component: OrderRating
-  },
-  {
-    path: '/ReviewAdmin',
-    name: 'ReviewAdmin',
-    component: ReviewsAdmin
-  },
-  {
-    path: '/ResturantReviews',
-    name: 'ResturantReviews',
-    component: ResturantReviews
-  },
-  {
-    path: '/Restaurantcard',
-    name: 'RestaurantCard',
-    component: RestaurantCard
-  },
-  {
     path: '/admin-dashboard',
     name: 'AdminDashboard',
-    component: AdminDashboard,
-    children: [
-      {
-        path: '/viewallorders',
-        name: 'ViewAllOrdersAdm', // Change name to avoid duplicates
-        component: ViewAllOrdersAdm // This is fine if you need this route
-      },
-    ]
+    component: AdminDashboard
   },
- 
   {
     path: '/AdminFinanceDashboard',
     name: 'AdminFinanceDashboard',
@@ -104,14 +81,7 @@ const routes = [
   {
     path: '/RestaurantOwnerDashboard',
     name: 'RestaurantOwner_Dashboard',
-    component: RestaurantOwner_Dashboard,
-    children: [
-      {
-        path: '/Resturant-owner',
-        name: 'RestaurantOwner',
-        component: ResturantOwner
-      }
-    ]
+    component: RestaurantOwner_Dashboard
   },
   {
     path: '/dashboardresturantpage',
@@ -138,27 +108,30 @@ const routes = [
     name: 'ModalView',
     component: ModalView
   },
-  
   {
-    path: '/',
-    name: 'Main_landing', // Change name to avoid duplicates
-    component: Main_landing // This is fine if you need this route
+    path: '/viewallorders',
+    name: 'ViewAllOrdersAdm', 
+    component: ViewAllOrdersAdm 
+  },
+  {
+    path: '/mainlanding',
+    name: 'Main_landing',
+    component: Main_landing 
   },
   {
     path: '/businesslanding',
-    name: 'Business_Landing', // Change name to avoid duplicates
-    component: Business_Landing // This is fine if you need this route
+    name: 'Business_Landing', 
+    component: Business_Landing 
   },
   {
     path: '/ownermenu',
-    name: 'OwnerMenu', // Change name to avoid duplicates
-    component: () => import('../views/Menu/OwnerMenu.vue') // This is fine if you need this route
+    name: 'OwnerMenu', 
+    component: () => import('../views/Menu/OwnerMenu.vue') 
   },
   {
     path: '/restOwnerProfile',
     name: 'RestOwnerProfile',
     component: RestOwnerProfile
-
   },
   {
     path: '/OrderReorder',
@@ -166,12 +139,21 @@ const routes = [
     component: OrderReorder,
   },
   {
-    path: '/order-details/:id',  // Define dynamic route with 'id' parameter
+    path: '/order-details/:id',  
     name: 'PrevorderDetails',
     component: PrevorderDetails,
-    props: true, // Pass route params as props to the component
-  }
-
+    props: true, 
+  },
+  {
+    path: '/user-guide',
+    name: 'UserGuide', 
+    component: () => import('../components/Customer/UserGuide.vue') 
+  },
+  {
+    path: '/recipe-blog',
+    name: 'RecipeBlog', 
+    component: () => import('../views/Customer/RecipeBlog.vue') 
+  },
 ]
 
 const router = createRouter({
