@@ -1,6 +1,8 @@
 <template>
     <div>
+        <!-- FilterAndSearch Component -->
         <FilterAndSearch @filterApplied="applyFilter" />
+        <!-- Table Component showing filteredOrders -->
         <Table :filteredOrders="filteredOrders" />
     </div>
 </template>
@@ -10,29 +12,39 @@ import { ref } from 'vue';
 import FilterAndSearch from '../../components/Admin/ViewAllOrdersFilterSearch.vue';
 import Table from '../../components/Admin/ViewAllOrdersTbl.vue';
 
-// Sample orders data (can be more complex with 100+ orders)
+// Orders data
 const orders = ref([
-    { id: 1, customerName: 'John Doe', restaurant: 'Restaurant A', date: '2024-10-01', status: 'In Progress', price: 30 },
-    { id: 2, customerName: 'Jane Smith', restaurant: 'Restaurant B', date: '2024-09-28', status: 'Completed', price: 50 },
-    { id: 3, customerName: 'Michael Brown', restaurant: 'Restaurant C', date: '2024-09-25', status: 'Cancelled', price: 20 },
-    { id: 4, customerName: 'Sarah Green', restaurant: 'Restaurant A', date: '2024-10-05', status: 'In Progress', price: 40 },
-    // Add more orders...
+    { id: 1, name: 'Ahmed Ali', phone: '0300-1234567', address: 'Karachi', resturant: "Cheezious", totalPrice: 1200, foodCommission: 50 },
+    { id: 2, name: 'Sara Khan', phone: '0312-9876543', address: 'Lahore', resturant: "KBJS", totalPrice: 1500, foodCommission: 70 },
+    { id: 3, name: 'Bilal Siddiqui', phone: '0345-5556677', address: 'Islamabad', resturant: "Ranchers", totalPrice: 1800, foodCommission: 80 },
+    { id: 4, name: 'Asma Sheikh', phone: '0301-2345678', address: 'Quetta', resturant: "Burger O Clock", totalPrice: 1400, foodCommission: 60 },
+    { id: 5, name: 'Hina Raza', phone: '0307-9988776', address: 'Peshawar', resturant: "Daily Deli", totalPrice: 2000, foodCommission: 90 },
+    { id: 6, name: 'Zainab Farooq', phone: '0322-4442233', address: 'Faisalabad', resturant: "Burger Lab", totalPrice: 1600, foodCommission: 75 },
+    { id: 7, name: 'Omer Malik', phone: '0302-1122334', address: 'Multan', resturant: "Howdy's", totalPrice: 1100, foodCommission: 55 }
 ]);
 
 const filteredOrders = ref([...orders.value]);
 
 // Apply filter logic
-// Apply filter logic
 const applyFilter = (filterData) => {
     filteredOrders.value = orders.value.filter((order) => {
-        const matchesStatus = filterData.selectedStatus === 'All' || order.status === filterData.selectedStatus;
-        const matchesPrice = order.price >= filterData.priceRange[0] && order.price <= filterData.priceRange[1];
-        const matchesStartDate = !filterData.startDate || new Date(order.date) >= new Date(filterData.startDate);
-        const matchesEndDate = !filterData.endDate || new Date(order.date) <= new Date(filterData.endDate);
-        const matchesRestaurant = !filterData.selectedRestaurant || order.restaurant === filterData.selectedRestaurant;
-        const matchesSearchQuery = !filterData.searchQuery || order.customerName.toLowerCase().includes(filterData.searchQuery.toLowerCase()) || order.restaurant.toLowerCase().includes(filterData.searchQuery.toLowerCase());
+        const matchesSearchQuery =
+            !filterData.searchQuery ||
+            order.name.toLowerCase().includes(filterData.searchQuery.toLowerCase()) ||
+            order.phone.includes(filterData.searchQuery) ||
+            order.address.toLowerCase().includes(filterData.searchQuery.toLowerCase());
 
-        return matchesStatus && matchesPrice && matchesStartDate && matchesEndDate && matchesRestaurant && matchesSearchQuery;
+        const matchesName = !filterData.nameFilter || order.name.toLowerCase().includes(filterData.nameFilter.toLowerCase());
+        const matchesPhone = !filterData.phoneFilter || order.phone.includes(filterData.phoneFilter);
+        const matchesAddress = !filterData.addressFilter || order.address.toLowerCase().includes(filterData.addressFilter.toLowerCase());
+
+        const matchesTotalPrice =
+            order.totalPrice >= filterData.priceRange[0] && order.totalPrice <= filterData.priceRange[1];
+        
+        const matchesCommission =
+            order.foodCommission >= filterData.commissionRange[0] && order.foodCommission <= filterData.commissionRange[1];
+
+        return matchesSearchQuery && matchesName && matchesPhone && matchesAddress && matchesTotalPrice && matchesCommission;
     });
 };
 </script>
