@@ -1,9 +1,12 @@
 <!-- Main_Landing.vue -->
 <template>
-  <BeforeLoginHeader @open-login="showLoginModal = true" @open-signup="showSignUpModal = true" />
+
   <div class="main d-flex">
+    <!-- Left half for content -->
     <div class="content-wrapper d-flex flex-column justify-content-center align-items-start">
-      <h2 class="text-left mb-3">It's the food and groceries you love, delivered</h2>
+      <h2 class="text-left mb-3">
+        It's the food and groceries you love, delivered
+      </h2>
 
       <div class="button-group mb-5">
         <button class="btn btn-login me-3" @click="showLoginModal = true">Login</button>
@@ -12,54 +15,33 @@
     </div>
 
 
-
-      <!-- Right half for image -->
-      <div class="image-wrapper">
-        <img src="../../assets/bg3.gif" alt="Landing Logo" class="logo" />
-      </div>
-
-      <LoginModal :showModal="showLoginModal" @close="closeLoginModal" />
-      <SignUpModal :showModal="showSignUpModal" @close="closeSignUpModal" />
-
+    <!-- Right half for image -->
+    <div class="image-wrapper">
+      <img src="../../assets/bg3.gif" alt="Landing Logo" class="logo" />
     </div>
 
-    
+    <LoginModal :showModal="showLoginModal" @close="closeLoginModal" />
+    <SignUpModal :showModal="showSignUpModal" @close="closeSignUpModal" />
+
+  </div>
 </template>
 
-<script>
-import { defineComponent, ref } from 'vue';
+<script setup>
+import { ref } from 'vue';
 import LoginModal from '../../components/LoginSignup/User_login.vue';
 import SignUpModal from '../../components/LoginSignup/User_signup.vue';
-import BeforeLoginHeader from '../../components/HeaderFooter/BeforeLoginHeader.vue';
 
+const showLoginModal = ref(false);
+const showSignUpModal = ref(false);
 
-export default defineComponent({
-  name: 'Main_Landing',
-  components: {
-    LoginModal,
-    SignUpModal,
-    BeforeLoginHeader,
-  },
-  setup() {
-    const showLoginModal = ref(false);
-    const showSignUpModal = ref(false);
+const closeLoginModal = () => {
+  showLoginModal.value = false;
+};
 
-    const closeLoginModal = () => {
-      showLoginModal.value = false;
-    };
+const closeSignUpModal = () => {
+  showSignUpModal.value = false;
+};
 
-    const closeSignUpModal = () => {
-      showSignUpModal.value = false;
-    };
-
-    return {
-      showLoginModal,
-      showSignUpModal,
-      closeLoginModal,
-      closeSignUpModal,
-    };
-  },
-});
 </script>
 
 <style scoped>
@@ -122,8 +104,9 @@ h2 {
 .logo {
   width: 100%;
   height: 100%;
-  object-fit: cover;
-  /* Ensures the image covers the entire right half */
+
+  object-fit: cover; /* Ensures the image covers the entire right half */
+
 }
 
 /* Responsive adjustments */
@@ -140,6 +123,7 @@ h2 {
 @media (max-width: 768px) {
   h2 {
     font-size: 1.8rem;
+
   }
 
   .button-group button {
@@ -152,8 +136,8 @@ h2 {
     flex-direction: column;
   }
 
-  .content-wrapper,
-  .image-wrapper {
+  .content-wrapper, .image-wrapper {
+
     width: 100%;
     height: 50vh;
   }
@@ -167,4 +151,6 @@ h2 {
     font-size: 0.8rem;
   }
 }
+
 </style>
+
